+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## 0.7.0
+## 0.6.3
+- Patch exports of async functions
+- Fix streaming unzip
+## 0.6.2
 - Replace Adler-32 implementation (used in Zlib compression) with one more optimized for V8
   - Advice from @SheetJSDev
 - Add support for extra fields, file comments in ZIP files